// src/app/website/model/actions.class.ts
export enum Actions {
  ADDRESS_COMPLETED = "ADDRESS_COMPLETED",
  ADDRESS_INCOMPLETE = "ADDRESS_INCOMPLETE",
  CART_MOVE_TO_STEP = "CART_MOVE_TO_STEP",
  INIT_CART = "INIT_CART",
  ADD_ARTICLE = "ADD_ARTICLE",
  REMOVE_ARTICLE = "REMOVE_ARTICLE",
  REMOVE_ORDER = "REMOVE_ORDER",
  EMPTY_CART = "EMPTY_CART",
  SET_PROFILE = "INIT_PROFILE"
}